test: :white_check_mark: Añadi test

+++ Encargado.java
@@ -26,30 +26,39 @@ public class Encargado
         }
     }
     
-    public void eliminarProducto(String codigo, String razon) {
+    public String eliminarProducto(String codigo, String razon) {
         Producto producto = inventario.obtenerProducto(codigo);
-        Date fechaRetiro = producto.getFechaRetiro();
-        String razonRetiro = producto.getRazonRetiro(razon);
-        boolean eliminado = inventario.eliminarProducto(codigo);
-        if (eliminado) {
-            System.out.println("Producto eliminado correctamente");
-            System.out.println("Fecha de retiro: " + fechaRetiro);
-            System.out.println("Razon de retiro: " + razonRetiro);
+        if (producto != null){
+            Date fechaRetiro = producto.getFechaRetiro();
+            String razonRetiro = producto.getRazonRetiro(razon);
+            boolean eliminado = inventario.eliminarProducto(codigo);
+            if (eliminado) {
+                System.out.println("Producto eliminado correctamente");
+                System.out.println("Fecha de retiro: " + fechaRetiro);
+                System.out.println("Razon de retiro: " + razonRetiro);
+                return "Producto eliminado correctamente";
+            }else {
+                System.out.println("No se pudo eliminar el producto");
+                return "No se pudo eliminar el producto";
+            }
         }
         else {
             System.out.println("El producto con el código " + codigo + " no existe en el inventario.");
+            return "El producto con el código " + codigo + " no existe en el inventario.";
         }
     }
 
-    public void modificarProducto(String codigo, int nStock, String nuevaInformacion) {
+    public String modificarProducto(String codigo, int nStock, String nuevaInformacion) {
         Producto producto = inventario.obtenerProducto(codigo);
         if (producto != null) {
             producto.setStock(nStock);
             producto.setInformacion(nuevaInformacion);
             System.out.println("Producto modificado correctamente");
+            return "Producto modificado correctamente";
         }
         else {
             System.out.println("El producto con el código " + codigo + " no existe en el inventario.");
+            return "El producto con el código " + codigo + " no existe en el inventario.";
         }
     }
